refactor(ui): extract message-list selection state
